requirements.txt
Django==2.1.3
django-filter==2.1.0
djangorestframework==3.9.0
jsonfield==2.0.2
requests==2.21.0
